Adds auto scaling to READ_HDF4
Two keywords have been added to the READ_HDF4 function:
/SCALE_DATA and /ADD_FIRST. When the /SCALE_DATA key is set,
scale factor, add offset, and missing/fill value attributes
that are found in the attributes of a variable are applied.
Default is to multiply by the scale factor and then add the
offset to the data. If the /ADD_FIRST keyword is set, then
the add offset is subtracted from the data then the scale
factor is applied. This is the MODIS convention. Missing
values are replaced with NaN.

README file also added

v0.0.0