Cleans up RSS readers
For almost every 'offical' Remote Sensing Systems data reader procedure,
there is a function I have written to duplicate/improve the reader.
It has been made clear which version of the reader is which, with
_krw appened to the name of each reader that I have written.

If my reader does not seem to be working correctly, please submit
a bug and revert to using the 'offical' procedure in the mean time.

v0.0.6